backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
httpx==0.25.1
python-multipart==0.0.6
websockets==12.0
aiohttp==3.9.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
pyOpenSSL==23.2.0
dnspython==2.4.2
beautifulsoup4==4.12.2
lxml==4.9.3
psutil==5.9.8
requests==2.31.0
python-owasp-zap-v2.4==0.0.20 